fix(test): add support to ignore all living test when they are not in live mode

+++ servicebus/data-plane/azure-servicebus/src/test/java/com/microsoft/azure/servicebus/ClientSessionTests.java
@@ -108,6 +108,9 @@ public abstract class ClientSessionTests extends Tests
     
     @AfterClass
     public static void cleanupAfterAllTest() throws ExecutionException, InterruptedException, IOException {
+        if (managementClient == null) {
+            return;
+        }
         if(ClientSessionTests.entityNameCreatedForAllTests != null)
         {
             managementClient.deleteQueueAsync(ClientSessionTests.entityNameCreatedForAllTests).get();
